refactor(server): split project-router.ts into per-domain register modules
Big bet #2. project-router.ts had grown to 4090 lines / 108 handlers in a
single createProjectRouter closure. This splits the handler body into
seven per-domain modules, each exporting a register<Domain>Routes(deps)
function that registers its routes on the SHARED router in the exact same
order — so Express route-matching precedence (e.g. jobs-export before
/jobs/:id) is byte-identical and behaviour is fully preserved.

- project-router-helpers.ts — shared spec/agent-model helpers, the
  Request.projectCtx augmentation, and the ProjectRoutesDeps contract.
  The public helpers are re-exported from project-router.ts so existing
  importers (`from './project-router'`) keep working.
- project-router-{jobs,spending,chat,setup,tickets,terminals,settings}.ts
  — the domain handler groups, registered in original order.
- createProjectRouter shrinks to a 201-line shell: middleware +
  sub-router mounts + the seven register calls.

The one cross-domain closure helper (ticketPath) is lifted into deps;
section-local helpers (toCsv, csvEscape, require*Enabled, parseRect, …)
move with their owning section. Zero behaviour change: full server suite
(2876 tests) green, server coverage 82.4% lines (>=80% gate), typecheck
clean.
