+++ server/project-router-helpers.ts
@@ -1,0 +1,310 @@
+// Shared helpers and the deps contract for the project-router domain modules.
+// Extracted verbatim from project-router.ts when that monolith was split into
+// per-domain register functions. Behaviour-preserving — same exports, same code.
+
+import fs from 'fs'
+import path from 'path'
+import type { Request, Router } from 'express'
+import type { ProjectRegistry, ProjectContext } from './project-registry'
+import { CORE_PACKAGE_SPEC } from './core-package'
+import {
+  getProviderDefault,
+  isValidModelForProvider,
+  type SpecProvider,
+} from './spec-models'
+import { clampShortSummary } from './ticket-store'
+
+/** Shared dependencies handed to every domain register function. */
+export interface ProjectRoutesDeps {
+  router: Router
+  registry: ProjectRegistry
+  ctx: (req: Request) => ProjectContext
+  /** Resolve the per-project ticket-store path for a request (cross-domain). */
+  ticketPath: (req: Request) => string
+}
+
+const TERMINAL_PANEL_ENABLED = process.env.SPECRAILS_TERMINAL_PANEL !== 'false'
+
+// ─── YAML helpers ─────────────────────────────────────────────────────────────
+
+/**
+ * Serialize install config to YAML matching specrails-core's tui-installer format exactly.
+ */
+function serializeInstallConfigYaml(config: Record<string, unknown>): string {
+  const c = config as {
+    version?: number
+    provider?: string
+    tier?: string
+    agents?: { selected?: string[]; excluded?: string[] }
+    models?: { preset?: string; defaults?: { model?: string }; overrides?: Record<string, string> }
+    agent_teams?: boolean
+  }
+
+  const overrides = c.models?.overrides ?? {}
+  const overridesEntries = Object.entries(overrides)
+  const overridesYaml = overridesEntries.length > 0
+    ? '\n' + overridesEntries.map(([k, v]) => `    ${k}: ${v}`).join('\n')
+    : ' {}'
+
+  const lines = [
+    '# specrails install config — generated by Specrails',
+    `# Re-run: npx ${CORE_PACKAGE_SPEC} init  to regenerate`,
+    `version: ${c.version ?? 1}`,
+    `provider: ${c.provider ?? 'claude'}`,
+    `tier: ${c.tier ?? 'quick'}`,
+    `agents:`,
+    `  selected: [${(c.agents?.selected ?? []).join(', ')}]`,
+    `  excluded: [${(c.agents?.excluded ?? []).join(', ')}]`,
+    `models:`,
+    `  preset: ${c.models?.preset ?? 'balanced'}`,
+    `  defaults: { model: ${c.models?.defaults?.model ?? 'sonnet'} }`,
+    `  overrides:${overridesYaml}`,
+    `agent_teams: ${c.agent_teams ?? false}`,
+    '',
+  ]
+
+  return lines.join('\n')
+}
+
+// ─── Agent model helpers ──────────────────────────────────────────────────────
+
+const VALID_MODEL_ALIASES = ['sonnet', 'opus', 'haiku'] as const
+type ModelAlias = typeof VALID_MODEL_ALIASES[number]
+
+/**
+ * Read installed agents from `.claude/agents/*.md` (top-level only, no subdirs).
+ * Extracts the `model:` field from YAML frontmatter.
+ */
+function readAgentModels(projectPath: string): { name: string; model: string }[] {
+  const agentsDir = path.join(projectPath, '.claude', 'agents')
+  if (!fs.existsSync(agentsDir)) return []
+
+  let entries: string[]
+  try {
+    entries = fs.readdirSync(agentsDir)
+  } catch {
+    return []
+  }
+
+  const agents: { name: string; model: string }[] = []
+  for (const entry of entries) {
+    // Top-level .md files only — skip subdirs
+    if (!entry.endsWith('.md')) continue
+    const filePath = path.join(agentsDir, entry)
+    try {
+      const stat = fs.statSync(filePath)
+      if (!stat.isFile()) continue
+      const content = fs.readFileSync(filePath, 'utf-8')
+      // Extract model from YAML frontmatter between --- markers
+      const frontmatterMatch = content.match(/^---\r?\n([\s\S]*?)\r?\n---/)
+      if (!frontmatterMatch) continue
+      const frontmatter = frontmatterMatch[1]
+      const modelMatch = frontmatter.match(/^model:\s*(.+)$/m)
+      const model = modelMatch ? modelMatch[1].trim() : 'sonnet'
+      agents.push({ name: entry.slice(0, -3), model })
+    } catch {
+      // skip unreadable files
+    }
+  }
+  return agents
+}
+
+/**
+ * Read `.specrails/install-config.yaml` and patch the `model:` line in each
+ * `.claude/agents/*.md` frontmatter to match the config's defaults/overrides.
+ * No-op if the config file does not exist.
+ */
+function applyModelConfig(projectPath: string): void {
+  const configPath = path.join(projectPath, '.specrails', 'install-config.yaml')
+  if (!fs.existsSync(configPath)) return
+
+  let configText: string
+  try {
+    configText = fs.readFileSync(configPath, 'utf-8')
+  } catch {
+    return
+  }
+
+  // Parse defaults.model
+  const defaultsMatch = configText.match(/defaults:\s*\{\s*model:\s*(\S+)\s*\}/)
+  const defaultModel: string = defaultsMatch ? defaultsMatch[1] : 'sonnet'
+
+  // Parse overrides block — lines like `    agentname: alias`
+  const overrides: Record<string, string> = {}
+  const overridesBlockMatch = configText.match(/overrides:([\s\S]*?)(?:\n\S|$)/)
+  if (overridesBlockMatch) {
+    const block = overridesBlockMatch[1]
+    const overrideLines = block.match(/^ {2,}(\S+):\s*(\S+)/gm) ?? []
+    for (const line of overrideLines) {
+      const m = line.match(/^\s+(\S+):\s*(\S+)/)
+      if (m) overrides[m[1]] = m[2]
+    }
+  }
+
+  const agentsDir = path.join(projectPath, '.claude', 'agents')
+  if (!fs.existsSync(agentsDir)) return
+
+  let entries: string[]
+  try {
+    entries = fs.readdirSync(agentsDir)
+  } catch {
+    return
+  }
+
+  for (const entry of entries) {
+    if (!entry.endsWith('.md')) continue
+    const filePath = path.join(agentsDir, entry)
+    try {
+      const stat = fs.statSync(filePath)
+      if (!stat.isFile()) continue
+      const agentName = entry.slice(0, -3)
+      const targetModel = overrides[agentName] ?? defaultModel
+      const content = fs.readFileSync(filePath, 'utf-8')
+      const patched = content.replace(/^model: .+$/m, `model: ${targetModel}`)
+      if (patched !== content) {
+        fs.writeFileSync(filePath, patched, 'utf-8')
+      }
+    } catch {
+      // skip unwritable files
+    }
+  }
+}
+
+/**
+ * Strip the metadata sections (Spec Title, Labels, Estimated Complexity,
+ * Short Summary) from a generate-spec LLM response so they don't end up
+ * duplicated inside the ticket's description — they're already parsed into
+ * dedicated fields and re-rendered by the UI.
+ */
+export function stripSpecMetadataSections(buffer: string): string {
+  return buffer
+    .replace(/##\s*Spec Title\s*\n+[^\n]*\n*/i, '')
+    .replace(/##\s*Labels\s*\n+(?:[^\n]+(?:\n(?!##)[^\n]+)*)\n*/i, '')
+    .replace(/##\s*Estimated Complexity\s*\n+(?:[^\n]+(?:\n(?!##)[^\n]+)*)\n*/i, '')
+    .replace(/##\s*Short Summary\s*\n+(?:[^\n]+(?:\n(?!##)[^\n]+)*)\n*/i, '')
+    .trim()
+}
+
+/**
+ * Extract the `## Short Summary` section body from a generate-spec response.
+ * Returns the raw multi-line string (the ticket-store `clampShortSummary`
+ * helper applies trim, control-strip, and the 240-char hard cap before
+ * persistence). Returns `null` when the section is missing or empty.
+ */
+export function extractShortSummary(buffer: string): string | null {
+  const m = buffer.match(/##\s*Short Summary\s*\n+((?:(?!##)[^\n]+(?:\n(?!##)[^\n]+)*))/i)
+  if (!m) return null
+  const body = m[1].trim()
+  return body.length > 0 ? body : null
+}
+
+export function deriveFallbackShortSummary(title: string, description: string): string | null {
+  const plain = description
+    .replace(/```[\s\S]*?```/g, ' ')
+    .replace(/^#{1,6}\s+.*$/gm, ' ')
+    .replace(/^\s*[-*]\s+/gm, '')
+    .replace(/\[[^\]]+\]\([^)]+\)/g, (m) => m.match(/\[([^\]]+)\]/)?.[1] ?? '')
+    .replace(/[`*_>]/g, '')
+    .replace(/\s+/g, ' ')
+    .trim()
+  const source = plain || title.trim()
+  if (!source) return null
+  const sentence = source.match(/^(.{24,}?[.!?])(?:\s|$)/)?.[1] ?? source
+  const capped = sentence.length > 160 ? `${sentence.slice(0, 157).trimEnd()}...` : sentence
+  return clampShortSummary(capped)
+}
+
+/**
+ * Lightly structure a raw free-form prompt (the "Raw" Add-Spec mode) when the
+ * user opts in. We deliberately do NOT fabricate spec sections — a raw prompt
+ * stays the user's own words. The only transform: when the body has no leading
+ * markdown heading, prefix a single neutral `## Overview` heading so the spec
+ * renders with a section title downstream. Returns the (trimmed) input
+ * unchanged when it already starts with a heading or is empty.
+ */
+export function lightlyStructurePrompt(text: string): string {
+  const trimmed = text.trim()
+  if (!trimmed) return trimmed
+  if (/^#{1,6}\s+/.test(trimmed)) return trimmed
+  return `## Overview\n\n${trimmed}`
+}
+
+/**
+ * Fold an `acceptanceCriteria` array into a ticket description body, writing
+ * (or replacing) a `## Acceptance Criteria` section.
+ *
+ * - `criteria.length > 0` → append/replace the section with one `- bullet` per item
+ * - `criteria.length === 0` → strip any existing `## Acceptance Criteria` section
+ *
+ * The match is case-insensitive on the heading text but requires `##` exactly
+ * to avoid matching other heading levels.
+ *
+ * Shared by `POST /tickets/from-draft` and `PATCH /tickets/:id`. See
+ * openspec/changes/replace-ai-edit-with-continue-editing/design.md D3+D4.
+ */
+export function formatDescriptionWithCriteria(body: string, criteria: string[]): string {
+  const sectionRegex = /\n*##\s*Acceptance Criteria\s*\n[\s\S]*?(?=\n##\s|\n*$)/i
+  const withoutExisting = body.replace(sectionRegex, '').replace(/\s+$/, '')
+  if (criteria.length === 0) return withoutExisting
+  const section = `## Acceptance Criteria\n\n${criteria.map((c) => `- ${c}`).join('\n')}`
+  if (withoutExisting === '') return section
+  return `${withoutExisting}\n\n${section}`
+}
+
+/**
+ * Resolve the default model used by Add Spec for a project.
+ *
+ * Order:
+ *   1. `models.defaults.model` from `<project>/.specrails/install-config.yaml`,
+ *      if it parses AND is in the provider allow-list.
+ *   2. Provider default from `PROVIDER_DEFAULT_MODEL` (`sonnet` / `gpt-5.5`).
+ *
+ * Logs a warning when the configured value exists but is not valid for the
+ * project's provider.
+ */
+export function resolveDefaultSpecModel(args: {
+  projectPath: string
+  provider: SpecProvider
+}): string {
+  const { projectPath, provider } = args
+  const configPath = path.join(projectPath, '.specrails', 'install-config.yaml')
+  if (!fs.existsSync(configPath)) return getProviderDefault(provider)
+
+  let configText: string
+  try {
+    configText = fs.readFileSync(configPath, 'utf-8')
+  } catch {
+    return getProviderDefault(provider)
+  }
+
+  const defaultsMatch = configText.match(/defaults:\s*\{\s*model:\s*(\S+?)\s*\}/)
+  const configured = defaultsMatch ? defaultsMatch[1] : null
+  if (!configured) return getProviderDefault(provider)
+
+  if (!isValidModelForProvider(configured, provider)) {
+    console.warn(
+      `[project-router] resolveDefaultSpecModel: configured model "${configured}" is not valid for provider "${provider}" — falling back to provider default`,
+    )
+    return getProviderDefault(provider)
+  }
+
+  return configured
+}
+
+// Extend Express Request to carry resolved ProjectContext
+declare module 'express-serve-static-core' {
+  interface Request {
+    projectCtx?: ProjectContext
+  }
+}
+
+// Re-export the locally-declared (non-exported) helpers so the domain modules
+// can import them.
+export {
+  TERMINAL_PANEL_ENABLED,
+  VALID_MODEL_ALIASES,
+  readAgentModels,
+  applyModelConfig,
+  serializeInstallConfigYaml,
+}
+export type { ModelAlias }
